services/api-gateway/requirements-test.txt
# Testing dependencies for API Gateway

# Core testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0

# Test fixtures and utilities
faker>=19.0.0
factory-boy>=3.3.0

# Mock and patching
responses>=0.23.0
freezegun>=1.2.2

# Redis testing
fakeredis>=2.19.0

# Flask testing
Flask-Testing>=0.8.1

# Code quality
coverage>=7.3.0
